+++ modules/EngineLayer/QueryManager/Helpers.h
@@ -104,8 +104,8 @@
   // returns NULL if dst_len is too short, otherwise return dst
   inline const char *ToHexString( char *dst, size_t dst_len, const char *src, const size_t src_len) {
     if ( !dst) return NULL;
-    int isrc = 0;
-    for ( int idst = 0; 
+    size_t isrc = 0;
+    for ( size_t idst = 0; 
 	  ( isrc < src_len) && ( idst < dst_len - 1); 
 	  isrc++, idst += 2) {
       sprintf( &dst[ idst], "%02x", ( unsigned char)src[ isrc]);
@@ -117,8 +117,8 @@
   inline const char *FromHexString( char *dst, size_t dst_len, const char *src, const size_t src_len) {
     if ( !dst) return NULL;
     bool error = false;
-    int isrc = 0;
-    for ( int idst = 0; 
+    size_t isrc = 0;
+    for ( size_t idst = 0; 
 	  ( isrc < src_len) && ( idst < dst_len); 
 	  isrc += 2, idst++) {
       int hi4 = tolower( src[ isrc]);
